Adequaçoes estruturais nas novas views do backend
